# backend/openapi/openapi.json
{"openapi":"3.1.0","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"Notion Integration","description":"Notion API integration endpoints for retrieving page and table content"},{"name":"Daily Report Management","description":"Daily report CRUD operations and export functionality"},{"name":"ToggleTrack Integration","description":"ToggleTrack API integration endpoints for retrieving time tracking data"},{"name":"AI Report Generation","description":"AI-powered daily report generation using GPT-5-mini"},{"name":"GitHub Integration","description":"GitHub MCP integration endpoints for daily report generation"}],"paths":{"/api/reports/{id}":{"put":{"tags":["Daily Report Management"],"summary":"Update daily report","description":"Update an existing daily report content","operationId":"updateReport","parameters":[{"name":"id","in":"path","description":"Report ID","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DailyReportUpdateRequestDto"}}},"required":true},"responses":{"404":{"description":"Report not found","content":{"application/json":{}}},"200":{"description":"Report updated successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DailyReportDto"}}}},"400":{"description":"Invalid request data or report not editable","content":{"application/json":{}}},"500":{"description":"Internal server error","content":{"application/json":{}}}}},"delete":{"tags":["Daily Report Management"],"summary":"Delete daily report","description":"Delete a daily report","operationId":"deleteReport","parameters":[{"name":"id","in":"path","description":"Report ID","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"404":{"description":"Report not found"},"500":{"description":"Internal server error","content":{"application/json":{}}},"204":{"description":"Report deleted successfully"}}}},"/api/reports":{"get":{"tags":["Daily Report Management"],"summary":"Get daily reports list","description":"Retrieve daily reports with optional filtering by date range and status","operationId":"getReports","parameters":[{"name":"userId","in":"query","description":"User ID","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"startDate","in":"query","description":"Start date for filtering (YYYY-MM-DD format)","required":false,"schema":{"type":"string","format":"date"},"example":"2024-01-01"},{"name":"endDate","in":"query","description":"End date for filtering (YYYY-MM-DD format)","required":false,"schema":{"type":"string","format":"date"},"example":"2024-12-31"},{"name":"status","in":"query","description":"Status filter","required":false,"schema":{"enum":["DRAFT","EDITED","APPROVED"]}}],"responses":{"400":{"description":"Invalid request parameters","content":{"application/json":{}}},"500":{"description":"Internal server error","content":{"application/json":{}}},"200":{"description":"Reports retrieved successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DailyReportListResponseDto"}}}}}},"post":{"tags":["Daily Report Management"],"summary":"Create daily report","description":"Create a new daily report","operationId":"createReport","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DailyReportCreateRequestDto"}}},"required":true},"responses":{"201":{"description":"Report created successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DailyReportDto"}}}},"500":{"description":"Internal server error","content":{"application/json":{}}},"400":{"description":"Invalid request data or report already exists for the date","content":{"application/json":{}}}}}},"/api/reports/{id}/regenerate":{"post":{"tags":["AI Report Generation"],"summary":"Regenerate daily report","description":"Regenerate an existing daily report with user feedback and additional information","operationId":"regenerateReport","parameters":[{"name":"id","in":"path","description":"Report ID","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReportRegenerationRequestDto"}}},"required":true},"responses":{"200":{"description":"Report regenerated successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReportGenerationResponseDto"}}}},"404":{"description":"Report not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReportGenerationResponseDto"}}}},"400":{"description":"Invalid request data","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReportGenerationResponseDto"}}}},"500":{"description":"Internal server error or AI regeneration failure","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReportGenerationResponseDto"}}}}}}},"/api/reports/{id}/approve":{"post":{"tags":["Daily Report Management"],"summary":"Approve daily report","description":"Approve a daily report, finalizing its content","operationId":"approveReport","parameters":[{"name":"id","in":"path","description":"Report ID","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"404":{"description":"Report not found","content":{"application/json":{}}},"200":{"description":"Report approved successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DailyReportDto"}}}},"500":{"description":"Internal server error","content":{"application/json":{}}}}}},"/api/reports/generate":{"post":{"tags":["AI Report Generation"],"summary":"Generate daily report","description":"Generate a new daily report using AI based on GitHub, Toggl, and Notion data","operationId":"generateReport","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReportGenerationRequestDto"}}},"required":true},"responses":{"500":{"description":"Internal server error or AI generation failure","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReportGenerationResponseDto"}}}},"201":{"description":"Report generated successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReportGenerationResponseDto"}}}},"400":{"description":"Invalid request data or report already exists for the date","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReportGenerationResponseDto"}}}}}}},"/api/reports/{id}/export":{"get":{"tags":["Daily Report Management"],"summary":"Export report as Markdown","description":"Export a daily report in Markdown format for download","operationId":"exportReport","parameters":[{"name":"id","in":"path","description":"Report ID","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"userName","in":"query","description":"User name for the export header","required":false,"schema":{"type":"string"}}],"responses":{"404":{"description":"Report not found","content":{"application/json":{}}},"500":{"description":"Internal server error","content":{"application/json":{}}},"200":{"description":"Markdown export successful","content":{"text/markdown":{"schema":{"format":"binary"}}}}}}},"/api/reports/{date}":{"get":{"tags":["Daily Report Management"],"summary":"Get daily report by date","description":"Retrieve a specific daily report for the given date","operationId":"getReportByDate","parameters":[{"name":"date","in":"path","description":"Report date in YYYY-MM-DD format","required":true,"schema":{"type":"string","format":"date"},"example":"2024-01-15"},{"name":"userId","in":"query","description":"User ID","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"400":{"description":"Invalid date format","content":{"application/json":{}}},"500":{"description":"Internal server error","content":{"application/json":{}}},"404":{"description":"Report not found for the specified date","content":{"application/json":{}}},"200":{"description":"Report retrieved successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DailyReportDto"}}}}}}},"/api/external/toggl/today":{"get":{"tags":["ToggleTrack Integration"],"summary":"Get today's time entries","description":"Retrieve all time entries for today from ToggleTrack. Backend automatically uses current date.","operationId":"getTodayTimeEntries","responses":{"500":{"description":"Internal server error or ToggleTrack API connection failure","content":{"application/json":{}}},"401":{"description":"Unauthorized - invalid ToggleTrack credentials","content":{"application/json":{}}},"200":{"description":"Today's time entries retrieved successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TimeEntryDto"}}}}}}},"/api/external/toggl/test":{"get":{"tags":["ToggleTrack Integration"],"summary":"Test ToggleTrack connection","description":"Test ToggleTrack API connection to verify access permissions","operationId":"testConnection","responses":{"500":{"description":"Internal server error or ToggleTrack API connection failure","content":{"application/json":{}}},"200":{"description":"Connection test completed successfully","content":{"application/json":{"schema":{"example":true}}}}}}},"/api/external/notion/test":{"get":{"tags":["Notion Integration"],"summary":"Test Notion connection","description":"Test Notion API connection to verify access permissions","operationId":"testConnection_1","responses":{"500":{"description":"Internal server error or Notion API connection failure","content":{"application/json":{}}},"200":{"description":"Connection test completed successfully","content":{"application/json":{"schema":{"example":true}}}}}}},"/api/external/notion/table":{"get":{"tags":["Notion Integration"],"summary":"Get table content","description":"Retrieve all rows from the pre-configured Notion table","operationId":"getTableContent","responses":{"500":{"description":"Internal server error or Notion API connection failure","content":{"application/json":{}}},"404":{"description":"Table not found or access denied","content":{"application/json":{}}},"200":{"description":"Table content retrieved successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/NotionTableRowDto"}}}}}}},"/api/external/notion/page":{"get":{"tags":["Notion Integration"],"summary":"Get page content","description":"Retrieve content from the pre-configured Notion page","operationId":"getPageContent","responses":{"500":{"description":"Internal server error or Notion API connection failure","content":{"application/json":{}}},"404":{"description":"Page not found or access denied","content":{"application/json":{}}},"200":{"description":"Page content retrieved successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/NotionPageDto"}}}}}}},"/api/external/github/test":{"get":{"tags":["GitHub Integration"],"summary":"Test GitHub connection","description":"Test GitHub repository connection via MCP to verify access permissions and repository existence","operationId":"testConnection_2","parameters":[{"name":"owner","in":"query","description":"Repository owner/organization name","required":true,"schema":{"type":"string"},"example":"octocat"},{"name":"repo","in":"query","description":"Repository name","required":true,"schema":{"type":"string"},"example":"Hello-World"}],"responses":{"400":{"description":"Invalid request parameters (missing owner or repo)","content":{"application/json":{}}},"200":{"description":"Connection test completed successfully","content":{"application/json":{"schema":{"example":true}}}},"500":{"description":"Internal server error or MCP connection failure","content":{"application/json":{}}}}}},"/api/external/github/pull-requests":{"get":{"tags":["GitHub Integration"],"summary":"Get pull requests for date","description":"Retrieve all pull requests created, updated, or merged on the specified date via MCP","operationId":"getPullRequestsForDate","parameters":[{"name":"owner","in":"query","description":"Repository owner/organization name","required":true,"schema":{"type":"string"},"example":"octocat"},{"name":"repo","in":"query","description":"Repository name","required":true,"schema":{"type":"string"},"example":"Hello-World"},{"name":"date","in":"query","description":"Target date in ISO format (YYYY-MM-DD)","required":true,"schema":{"type":"string","format":"date"},"example":"2024-01-15"}],"responses":{"200":{"description":"Pull requests retrieved successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PullRequestDto"}}}},"500":{"description":"Internal server error or MCP connection failure","content":{"application/json":{}}},"400":{"description":"Invalid request parameters or date format","content":{"application/json":{}}},"404":{"description":"Repository not found or access denied","content":{"application/json":{}}}}}},"/api/external/github/commits":{"get":{"tags":["GitHub Integration"],"summary":"Get commits for date","description":"Retrieve all commits made on the specified date via MCP","operationId":"getCommitsForDate","parameters":[{"name":"owner","in":"query","description":"Repository owner/organization name","required":true,"schema":{"type":"string"},"example":"octocat"},{"name":"repo","in":"query","description":"Repository name","required":true,"schema":{"type":"string"},"example":"Hello-World"},{"name":"date","in":"query","description":"Target date in ISO format (YYYY-MM-DD)","required":true,"schema":{"type":"string","format":"date"},"example":"2024-01-15"}],"responses":{"200":{"description":"Commits retrieved successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CommitDto"}}}},"500":{"description":"Internal server error or MCP connection failure","content":{"application/json":{}}},"400":{"description":"Invalid request parameters or date format","content":{"application/json":{}}},"404":{"description":"Repository not found or access denied","content":{"application/json":{}}}}}}},"components":{"schemas":{"DailyReportUpdateRequestDto":{"type":"object","description":"Report update request","properties":{"editedContent":{"type":"string"},"additionalNotes":{"type":"string"}}},"DailyReportDto":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"userId":{"type":"string","format":"uuid"},"reportDate":{"type":"string","format":"date"},"rawData":{"type":"string"},"generatedContent":{"type":"string"},"editedContent":{"type":"string"},"finalContent":{"type":"string"},"status":{"type":"string"},"generationCount":{"type":"integer","format":"int32"},"additionalNotes":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"displayContent":{"type":"string"}}},"DailyReportCreateRequestDto":{"type":"object","description":"Report creation request","properties":{"userId":{"type":"string","format":"uuid"},"reportDate":{"type":"string","format":"date"},"rawData":{"type":"string"},"generatedContent":{"type":"string"},"additionalNotes":{"type":"string"},"valid":{"type":"boolean"}}},"ReportRegenerationRequestDto":{"type":"object","description":"Report regeneration request","properties":{"reportId":{"type":"string","format":"uuid"},"userFeedback":{"type":"string"},"additionalNotes":{"type":"string"},"valid":{"type":"boolean"}}},"ReportGenerationResponseDto":{"type":"object","properties":{"reportId":{"type":"string","format":"uuid"},"userId":{"type":"string","format":"uuid"},"reportDate":{"type":"string","format":"date"},"generatedContent":{"type":"string"},"status":{"type":"string"},"generationCount":{"type":"integer","format":"int32"},"generatedAt":{"type":"string","format":"date-time"},"success":{"type":"boolean"},"errorMessage":{"type":"string"}}},"ReportGenerationRequestDto":{"type":"object","description":"Report generation request","properties":{"userId":{"type":"string","format":"uuid"},"reportDate":{"type":"string","format":"date"},"additionalNotes":{"type":"string"},"valid":{"type":"boolean"}}},"DailyReportListResponseDto":{"type":"object","properties":{"reports":{"type":"array","items":{"$ref":"#/components/schemas/DailyReportDto"}},"totalCount":{"type":"integer","format":"int32"},"dateRange":{"type":"string"},"status":{"type":"string"},"actualCount":{"type":"integer","format":"int32"}}},"TimeEntryDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"description":{"type":"string"},"projectName":{"type":"string"},"projectId":{"type":"integer","format":"int64"},"startTime":{"type":"string","format":"date-time"},"endTime":{"type":"string","format":"date-time"},"durationSeconds":{"type":"integer","format":"int64"},"billable":{"type":"boolean"},"userId":{"type":"string"},"workspaceName":{"type":"string"},"workspaceId":{"type":"integer","format":"int64"},"tags":{"type":"array","items":{"type":"string"}}}},"NotionTableRowDto":{"type":"object","properties":{"id":{"type":"string"},"title":{"type":"string"},"properties":{"type":"object","additionalProperties":{}},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"url":{"type":"string"}}},"NotionPageDto":{"type":"object","properties":{"id":{"type":"string"},"title":{"type":"string"},"content":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"url":{"type":"string"},"tags":{"type":"array","items":{"type":"string"}},"status":{"type":"string"}}},"PullRequestDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"number":{"type":"integer","format":"int32"},"title":{"type":"string"},"body":{"type":"string"},"state":{"type":"string"},"author":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"mergedAt":{"type":"string","format":"date-time"},"baseBranch":{"type":"string"},"headBranch":{"type":"string"},"reviewers":{"type":"array","items":{"type":"string"}},"additions":{"type":"integer","format":"int32"},"deletions":{"type":"integer","format":"int32"},"changedFiles":{"type":"integer","format":"int32"},"htmlUrl":{"type":"string"}}},"CommitDto":{"type":"object","properties":{"sha":{"type":"string"},"message":{"type":"string"},"author":{"type":"string"},"authorEmail":{"type":"string"},"date":{"type":"string","format":"date-time"},"additions":{"type":"integer","format":"int32"},"deletions":{"type":"integer","format":"int32"},"total":{"type":"integer","format":"int32"},"htmlUrl":{"type":"string"}}}}}}